requirements.txt
Flask==3.0.0
google-genai>=1.49.0
Werkzeug==3.0.1
python-dotenv==1.0.0
